Updated cosmos.py

+++ cosmos.py
@@ -88,6 +88,8 @@ async def shutup_bot(ctx):
 
 @bot.check
 async def tad_general_check(ctx):
+    if ctx.message.author.guild_permissions.manage_guild:
+        return True
     if ctx.message.channel.id == 244998983112458240:
         await ctx.message.add_reaction('👎')
         await ctx.message.delete()
